Restructure modal component dir
