+++ components/class-pg/WriteReportModal.tsx
@@ -8,7 +8,7 @@ import { useState } from "react";
 // import FormSubmitButton from "../FormSubmitButton";
 import Button from "../Button";
 
-import ModalOuter from "../modal-parent-components/ModalOuter";
+import ModalOuter from "../ModalOuter";
 
 import { ClassSubjectGroupStudent } from "@/types/types";
